+++ test/e2e/v3.fetch.spec.ts
@@ -158,4 +158,16 @@ describe('v3.fetch', () => {
             })
         );
     });
+
+    it('it should parse query params', async () => {
+        const result = await browser.evaluate(async () => {
+            const { ParametersService } = (window as any).api;
+            return (await ParametersService.postCallWithOptionalParam({
+                page: 0,
+                size: 1,
+                sort: ['location'],
+            })) as Promise<any>;
+        });
+        expect(result.query).toStrictEqual({ parameter: { page: '0', size: '1', sort: 'location' } });
+    });
 });