+++ test/e2e/v3.node.spec.ts
@@ -138,4 +138,14 @@ describe('v3.node', () => {
             })
         );
     });
+
+    it('it should parse query params', async () => {
+        const { ParametersService } = require('./generated/v3/node/index.js');
+        const result = (await ParametersService.postCallWithOptionalParam({
+            page: 0,
+            size: 1,
+            sort: ['location'],
+        })) as Promise<any>;
+        expect((result as any).query).toStrictEqual({ parameter: { page: '0', size: '1', sort: 'location' } });
+    });
 });